agregado metodo post

+++ RepasoGeneral/Models/Espacio.cs
@@ -8,6 +8,6 @@ public class Espacio
 {
     public int Id {get; set;}
     public string Nombre {get; set;}
-    public decimal PrecioHora {get; set;}
+    public double PrecioHora {get; set;}
     public List<Reserva> Reservas {get; set;}
 }
